Move Icons to icons folder, where it belongs
